remove pcre1 support
I never use the pcre1 library, so it's advantageous to remove a layer of
abstraction and support only pcre2 going forward.

Possibly the biggest advantage here is that every call to ag_pcre_match
doesn't have to allocate and free the pcre2_match_data structure and
copy the ovector data; the same match data struct can be re-used for
multiple matches.

In a benchmark searching for a URL regex in the linux kernel tree with
filetypes enabled, there's a 20% end-to-end speedup from 1.73s down to 1.38s.

./ag --stats -c --workers=1 --cc --asm '[\w]+://[^/\s?#]+[^\s?#]+(?:\?[^\s#]*)?(?:#[^\s]*)?' /tmp/linux-src

While I'm here, I cleaned up checking for JIT (use it for all regexes
rather than just the main one).

+++ src/print.c
@@ -100,7 +100,7 @@ void print_trailing_context(const char *path, const char *buf, size_t n) {
     }
 }
 
-void print_path_match(const char *_path, const char sep, int *match_ovec) {
+void print_path_match(const char *_path, const char sep, size_t *match_ovec) {
     if (opts.print_path == PATH_PRINT_NOTHING && !opts.vimgrep) {
         return;
     }
@@ -117,8 +117,8 @@ void print_path_match(const char *_path, const char sep, int *match_ovec) {
                 // This only works when using opts.match_files (-g option), not when combining
                 // -l and -G, which is otherwise equivalent but goes through the whole search_buf
                 // process and the path match details are lost by the time we get here
-                int ovec0 = match_ovec[0];
-                int ovec1 = match_ovec[1];
+                size_t ovec0 = match_ovec[0];
+                size_t ovec1 = match_ovec[1];
                 size_t orig_pathlen = strlen(_path);
                 size_t pathlen = strlen(path);
                 if (pathlen < orig_pathlen) {
